@@ -12,55 +12,58 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.0.0",
-    "@angular/cdk": "^5.0.2",
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/flex-layout": "^2.0.0-beta.12",
-    "@angular/forms": "^5.0.0",
-    "@angular/http": "^5.0.0",
-    "@angular/material": "^5.0.2",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@angular/router": "^5.0.0",
+    "@angular/animations": "^8.2.0",
+    "@angular/cdk": "^8.1.2",
+    "@angular/common": "^8.2.0",
+    "@angular/compiler": "^8.2.0",
+    "@angular/core": "^8.2.0",
+    "@angular/flex-layout": "^8.0.0-beta.26",
+    "@angular/forms": "^8.2.0",
+    "@angular/material": "^8.1.2",
+    "@angular/platform-browser": "^8.2.0",
+    "@angular/platform-browser-dynamic": "^8.2.0",
+    "@angular/router": "^8.2.0",
     "@ngrx/core": "^1.2.0",
-    "@ngrx/effects": "^4.1.1",
-    "@ngrx/store": "^4.1.1",
+    "@ngrx/effects": "^6.0.0",
+    "@ngrx/store": "^6.0.0",
     "angular-particle": "^1.0.4",
-    "angular-star-rating": "^3.0.8",
+    "angular-star-rating": "^4.0.0-beta.3",
     "animate.css": "^3.6.1",
-    "core-js": "^2.4.1",
+    "core-js": "^3.1.4",
     "font-awesome": "^4.7.0",
-    "hapi": "^17.2.2",
+    "hapi": "^18.1.0",
     "inert": "^5.1.0",
-    "material-design-icons-iconfont": "^3.0.3",
-    "ngrx-store-localstorage": "^0.3.0",
-    "ngx-quill": "^2.1.2",
-    "rxjs": "^5.5.2",
+    "material-design-icons-iconfont": "^5.0.1",
+    "ngrx-store-localstorage": "^8.0.0",
+    "ngx-quill": "^7.1.2",
+    "path": "^0.12.7",
+    "quill": "^1.3.6",
+    "rxjs": "^6.5.2",
     "swagger-ui-dist": "^3.10.0",
-    "zone.js": "^0.8.14"
+    "tslib": "^1.10.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular/cli": "^1.6.2",
-    "@angular/compiler-cli": "^5.0.0",
-    "@angular/language-service": "^5.0.0",
-    "@ngrx/store-devtools": "^4.1.1",
-    "@types/jasmine": "~2.5.53",
+    "@angular-devkit/build-angular": "~0.802.0",
+    "@angular/cli": "^8.2.0",
+    "@angular/compiler-cli": "^8.2.0",
+    "@angular/language-service": "^8.2.0",
+    "@ngrx/store-devtools": "^6.0.0",
+    "@types/jasmine": "~3.3.16",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~3.2.0",
-    "tslint": "~5.7.0",
-    "typescript": "~2.4.2"
+    "@types/node": "~12.6.8",
+    "codelyzer": "^5.0.1",
+    "jasmine-core": "~3.4.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~4.2.0",
+    "karma-chrome-launcher": "~3.0.0",
+    "karma-cli": "~2.0.0",
+    "karma-coverage-istanbul-reporter": "^2.1.0",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
+    "protractor": "~5.4.2",
+    "ts-node": "~8.3.0",
+    "tslint": "~5.18.0",
+    "typescript": "~3.5.3"
   }
 }
