@@ -4,7 +4,7 @@ import { LoginMenuAction, LoginResponseData, LoginMenuActionTypes, LoginFormData
 import { Store } from '@ngrx/store';
 import { AppState } from './app.data.models';
 import { DoLogoutAction } from './authentication/authentication.actions';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import * as loginActions from './authentication/authentication.actions';
 import * as globalActions from './app.actions';
 import { ToggleLeftPanelAction, ToggleParticleAction } from './app.actions';
@@ -53,7 +53,7 @@ export class AppComponent implements OnInit {
 
     setTimeout(() => {
       this.store.dispatch(new ToggleLeftPanelAction(true));
-      this.store.dispatch(new ToggleParticleAction(true));
+      this.store.dispatch(new ToggleParticleAction(false));
       this.store.dispatch(new globalActions.ToggleMenuBackgroundAction(false));
     }, 200);
 
