@@ -1,6 +1,10 @@
+
+import {filter} from 'rxjs/operators';
 import { Component, OnInit, Input } from '@angular/core';
 import { Router, NavigationEnd } from '@angular/router';
-import 'rxjs/add/operator/filter';
+import { Store } from '@ngrx/store';
+import { AppState } from '../../../app.data.models';
+import * as globalActions from '../../../app.actions';
 
 @Component({
   selector: 'store-breadcrumbs',
@@ -13,12 +17,12 @@ export class BreadcrumbsComponent implements OnInit {
   @Input()
   public isOpen: boolean;
 
-  constructor(private _router: Router) {
+  constructor(private _router: Router, private store: Store<AppState>) {
   }
 
   ngOnInit() {
-    this._router.events
-      .filter((event: any) => event instanceof NavigationEnd)
+    this._router.events.pipe(
+      filter((event: any) => event instanceof NavigationEnd))
       .subscribe((event: NavigationEnd) => {
         const tmp = event.url.replace('/', '').split('/');
         if (tmp[0] === 'legacy') {
@@ -29,5 +33,9 @@ export class BreadcrumbsComponent implements OnInit {
       });
 
   }
-}
+
+  onHambergerMenuClick(event) {
+    this.store.dispatch(new globalActions.ToggleLeftPanelAction(event));
+  }
+} 
 
